The Homepage
This page features a welcoming banner with attractive tropical background. This page also includes navigation links such as Home, About, Services, Shopping, contact and Enquiry, and User login account.

Moreover, there is a call to action button for exploring products which enhances user engagement on the website. The page also includes a small description of the services and products sold by the company