Products Categories page
This page displays a categorized products listing with their real images, description and prices.

For instance, the side bar menu allows for filtering by categories such as Fishing gears, Life jackets, Services and Solar.

on this page, there is also a clean and structured layout with a 'Quick Enquiry'  section where a client can simply enquire anything to the company by filling the form of his enquiry.